+++ Projekt Info/src/Inputs/KeyboardListener.java
@@ -1,0 +1,5 @@
+package Inputs;
+
+public interface KeyboardListener {
+	public void action(int EventKey);
+}
